feat(#715): move agreement aside pages
