+++ apps/web/src/app/[lang]/dho/[id]/agreements/@aside/create/loading.tsx
@@ -1,12 +1,10 @@
 'use client';
 
 import { CreateAgreementForm } from '@hypha-platform/epics';
-import { useParams } from 'next/navigation';
-import { SidePanel } from '../../_components/side-panel';
+import { SidePanel } from '../../../@aside/_components/side-panel';
 import { useMe } from '@hypha-platform/core/client';
 
-export default function CreateAgreement() {
-  const { lang, id } = useParams();
+export default function Loading() {
   const { person } = useMe();
   return (
     <SidePanel>
@@ -16,11 +14,11 @@ export default function CreateAgreement() {
           name: person?.name || '',
           surname: person?.surname || '',
         }}
-        closeUrl={`/${lang}/dho/${id}/agreements`}
+        closeUrl={``}
         onCreate={() => {
           console.log('Publish proposal');
         }}
-        isLoading={false}
+        isLoading={true}
       />
     </SidePanel>
   );
