+++ apps/web/src/app/[lang]/dho/[id]/agreements/@aside/page.tsx
@@ -1,0 +1,3 @@
+export default function Default() {
+  return null;
+}
