+++ apps/web/src/app/[lang]/dho/[id]/agreements/default.tsx
@@ -1,0 +1,2 @@
+import Page from './page';
+export default Page;
