+++ apps/web/src/app/[lang]/dho/[id]/agreements/layout.tsx
@@ -1,0 +1,20 @@
+export const metadata = {
+  title: 'Hypha | My Spaces',
+  description: 'Generated by create-nx-workspace',
+};
+
+export default async function RootLayout({
+  children,
+  aside,
+}: {
+  children: React.ReactNode;
+  aside: React.ReactNode;
+}) {
+  console.debug('MySpace/layout');
+  return (
+    <div className="w-full h-full flex">
+      {children}
+      {aside}
+    </div>
+  );
+}
